commands: Move the matrix command into a separate file.
